bpf: Avoid implicit shorten-64-to-32 in clang 19
As per the below PR, clang 19.1.x is having -Wshorten-64-to32 under the
-Wimplicit-int-conversion, which is already forbidden in Makefile.bpf.
This commit is to avoid the conversion error as per below sample

```
./lib/nodeport.h:1825:18: error: implicit conversion loses integer precision: '__u64' (aka 'unsigned long long') to '__s32' (aka 'int') [-Werror,-Wshorten-64-to-32]
 1825 |         __s32 len_old = ctx_full_len(ctx);
```

Relates: https://github.com/llvm/llvm-project/pull/80814

+++ bpf/include/bpf/ctx/xdp.h
@@ -178,8 +178,8 @@ l3_csum_replace(const struct xdp_md *ctx, __u64 off, const __u32 from,
 		       [offmax]"i"(__CTX_OFF_MAX), [errno]"i"(-EINVAL)
 		     : "r1", "r2");
 	if (!ret)
-		from ? __csum_replace_by_4(sum, from, to) :
-		       __csum_replace_by_diff(sum, to);
+		from ? __csum_replace_by_4(sum, (__u32)from, (__u32)to) :
+		       __csum_replace_by_diff(sum, (__u32)to);
 	return ret;
 }
 
@@ -386,7 +386,7 @@ ctx_full_len(const struct xdp_md *ctx)
 static __always_inline __maybe_unused __u32
 ctx_wire_len(const struct xdp_md *ctx)
 {
-	return ctx_full_len(ctx);
+	return (__u32)ctx_full_len(ctx);
 }
 
 struct {
